Dung @RestControllerAdvice de xu ly global Exception

+++ demo/src/main/java/com/example/demo/Controller/UserController.java
@@ -1,7 +1,10 @@
 package com.example.demo.Controller;
 
 import com.example.demo.Domain.User;
+import com.example.demo.Service.Error.IDInvalidException;
 import com.example.demo.Service.UserService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,38 +17,42 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/user/{id}")
-    public User fetchUserByID(@PathVariable("id") long id){
+    @GetMapping("/users/{id}")
+    public ResponseEntity<User> fetchUserByID(@PathVariable("id") long id){
         User fetchUser = this.userService.getUserByID(id);
-        return fetchUser;
+        return ResponseEntity.status(HttpStatus.OK).body(fetchUser);
     }
 
-    @GetMapping("/user")
-    public List<User> fetchAllUsers(){
-        return this.userService.getAllUsers();
+    @GetMapping("/users")
+    public ResponseEntity<List<User>> fetchAllUsers(){
+        return ResponseEntity.status(HttpStatus.OK).body(this.userService.getAllUsers());
     }
 
-    @PostMapping("/user")
-    public User createNewUser(
+    @PostMapping("/users")
+    public ResponseEntity<User> createNewUser(
             @RequestBody User postManUser) {
         User user = new User();
         user = this.userService.handleCreateUser(postManUser);
-        return user;
+        return ResponseEntity.status(HttpStatus.CREATED).body(user);
     }
 
-    @DeleteMapping("/user/{id}")
-    public String deleteUser(@PathVariable("id") long id) {
+    @DeleteMapping("/users/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable("id") long id) throws IDInvalidException {
+        if(id >= 1500){
+            throw new IDInvalidException("Id khong lon hon 1500");
+        }
         User deleteUser = this.userService.getUserByID(id);
         if (deleteUser != null) {
             this.userService.deleteUserByID(id);
         }
-        return "xóa thành công";
+        return ResponseEntity.status(HttpStatus.OK).body("Delete users successfully");
     }
 
-    @PutMapping("/user")
-    public User updateUser(
+    @PutMapping("/users")
+    public ResponseEntity<User> updateUser(
             @RequestBody User postManUser){
         User updateUser = this.userService.updateUser(postManUser);
-        return updateUser;
+        return ResponseEntity.status(HttpStatus.OK).body(updateUser);
     }
+
 }
